responding to all PR requests. start pose as arrray, cleaned up and renamed to test_leap_grasp.py
