+++ config/default_params.yaml
@@ -1,4 +1,4 @@
-baudrate: 4000000
+baudrate: 3000000
 curr_lim: 5000
 device_name: /dev/serial/by-id/usb-FTDI_USB__-__Serial_Converter_FT8ISZ8G-if00-port0
 hand_name: dom
@@ -8,20 +8,4 @@ kP: 600
 pub_current: false
 pub_pos: true
 pub_vel: false
-start_pos:
-- 0
-- 0
-- 0
-- 0
-- 0
-- 0
-- 0
-- 0
-- 0
-- 0
-- 0
-- 0
-- 0
-- 0
-- 0
-- 0
+start_pos: [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
